Task/fp 511 onboarding view unit tests (#78)
* staff_user fixture

* Refactor top level onboarding view tests

* Add note

* Refactor tests so they are testing route

* Remove print
