+++ server/portal/apps/onboarding/api/views_unit_test.py
@@ -30,25 +30,6 @@ def mocked_log_setup_state(mocker):
     yield mocker.patch('portal.apps.onboarding.api.views.log_setup_state')
 
 
-@pytest.fixture
-def mock_steps(regular_user, settings):
-    settings.PORTAL_USER_ACCOUNT_SETUP_STEPS = ['portal.apps.onboarding.steps.test_steps.MockStep']
-    pending_step = SetupEvent.objects.create(
-        user=regular_user,
-        step="portal.apps.onboarding.steps.test_steps.MockStep",
-        state=SetupState.PENDING,
-        message="message"
-    ).save()
-
-    completed_step = SetupEvent.objects.create(
-        user=regular_user,
-        step="portal.apps.onboarding.steps.test_steps.MockStep",
-        state=SetupState.COMPLETED,
-        message="message",
-    ).save()
-    yield (pending_step, completed_step,)
-
-
 """
 SetupStepView tests
 """
